Ownership for Lab Sessions.
Added a new field for ownership to the Lab Session class.
Listing Lab Sessions depending on owner and not on project.
Removal of Lab Sessions is not possible unless an owner even when the project is shared between users.

Restricted 'Grant access to project' so that only the owner of the project can share it.

More info -> [JIRA-13](https://scaleoutsystems.atlassian.net/browse/STACKN-13) , [https://scaleoutsystems.atlassian.net/browse/STACKN-28)